Use one verif_global_util.py
